% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_add_LCT.R
\name{bib_add_LCT}
\alias{bib_add_LCT}
\title{This function removes entries with the same key in a .bib file and writes a new .bib file}
\usage{
bib_add_LCT(tib_doc)
}
\arguments{
\item{tib_doc}{a tib_doc tibble}
}
\value{
a tib_doc tibble with added column LCT (locally cited times)
}
\description{
This function removes entries with the same key in a .bib file and writes a new .bib file
}
