% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_remove_duplicates.R
\name{bib_remove_duplicates}
\alias{bib_remove_duplicates}
\title{This function removes entries with the same key in a .bib file and writes a new .bib file}
\usage{
bib_remove_duplicates(from_file, to_file = "savedrecs_clean.bib")
}
\arguments{
\item{from_file}{the path to the .bib file}

\item{to_file}{the file with no duplicates created}
}
\value{
nothing
}
\description{
This function removes entries with the same key in a .bib file and writes a new .bib file
}
