% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_tib_doc.R
\name{bib_tib_doc}
\alias{bib_tib_doc}
\title{This function reads bib entries from a .bib file and formats them for later use with package bibou}
\usage{
bib_tib_doc(bibtex_file, check_in_text = NULL)
}
\arguments{
\item{bibtex_file}{the path to the .bib file}

\item{check_in_text}{filter docs to keep only those with this pattern in either title, keywords, or abstract. Defaults to NULL.}
}
\value{
a tibble with one row = one document
}
\description{
This function reads bib entries from a .bib file and formats them for later use with package bibou
}
\examples{
bibtex_file=system.file("savedrecs_Rhone.bib",package="bibou")
bib_tib_doc(bibtex_file)
}
