% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_tib_doc_ABW.R
\name{bib_tib_doc_ABW}
\alias{bib_tib_doc_ABW}
\title{This function reads bib entries from a .bib file and formats them for later use with package bibou}
\usage{
bib_tib_doc_ABW(tib_doc)
}
\arguments{
\item{tib_doc}{a tibble with one row=one document}
}
\value{
a tibble with one row = one document*abstract word
}
\description{
This function reads bib entries from a .bib file and formats them for later use with package bibou
}
\examples{
data(tib_doc_Rhone)
bib_tib_doc_ABW(tib_doc_Rhone)
}
