% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_tib_doc_AU.R
\name{bib_tib_doc_AU}
\alias{bib_tib_doc_AU}
\title{This function reads bib entries from a .bib file and formats them for later use with package bibou}
\usage{
bib_tib_doc_AU(tib_doc)
}
\arguments{
\item{tib_doc}{a tibble with one row=one document}
}
\value{
a tibble with one row = one document*author
}
\description{
This function reads bib entries from a .bib file and formats them for later use with package bibou
}
\examples{
bib_tib_doc_author(tib_doc_Rhone)
}
