% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_tib_doc_keyword.R
\name{bib_tib_doc_keyword}
\alias{bib_tib_doc_keyword}
\title{This function reads bib entries from a .bib file and formats them for later use with package bibou}
\usage{
bib_tib_doc_keyword(tib_doc)
}
\arguments{
\item{tib_doc}{a tibble with one row=one document}
}
\value{
a tibble with one row = one document*keyword
}
\description{
This function reads bib entries from a .bib file and formats them for later use with package bibou
}
\examples{
bibtex_file=system.file("savedrecs_Rhone.bib",package="bibou")
tib_doc=bib_tib_doc(bibtex_file)
bib_tib_doc_keyword(tib_doc)
}
