% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_remove_duplicates.R
\name{bib_remove_duplicates}
\alias{bib_remove_duplicates}
\title{This function removes entries with the same key in a .bib file and writes a new .bib file}
\usage{
bib_remove_duplicates(bibtex_file, suffix = "_clean")
}
\arguments{
\item{bibtex_file}{the path to the .bib file}

\item{suffix}{the suffix to use to create new .bib file (the new file name will be bibtex_file_\{suffix\}). Defaults to "_clean"}
}
\value{
nothing
}
\description{
This function removes entries with the same key in a .bib file and writes a new .bib file
}
